requirements.txt
joblib==1.0.0
validators==0.18.2
pandas==1.1.5
scikit_learn==0.24.1
PyYAML==5.4.1